Organize icon components
Move icon components from `src/components/logos` into a dedicated
`src/components/icons` directory to better separate logos from
general icons.

Update imports to reflect the new location.

Remove the `/icons` route and the `AlternateLogos` component, which
were used for previewing the icons.
