+++ src/components/icons/OnetimeSecretIcon.vue
