+++ src/components/icons/sprites/AlternateLogoSprites.vue
