+++ src/components/logos/DefaultLogo.vue
@@ -3,7 +3,7 @@
 <script setup lang="ts">
   import { computed } from 'vue';
   import { useI18n } from 'vue-i18n';
-  import MonotoneJapaneseSecretButton from './MonotoneJapaneseSecretButton.vue';
+  import MonotoneJapaneseSecretButton from '@/components/icons/MonotoneJapaneseSecretButtonIcon.vue';
 
 import { type LogoConfig } from '@/types/ui/layouts';
